refactor(logging): update CLI logger setup and tests for improved clarity and functionality
